Fix remaining 121 compilation errors - build now successful
Major fixes:
- UIDialogs: Added vararg Descriptor constructor, multiShowDialog overloads,
  showDialogProgress with ProgressHandler interface, overloaded showGeneralErrorDialog
- Settings.kt: Complete stub with all required properties (autoplay, broadcast,
  localConnections, connectLast, pairThroughRelay, discoverThroughRelay, peekChannelContents,
  checkDisabledPluginsForUpdates, watchLaterAddStart, playlistAllowDups, polycentricEnabled),
  encode()/replace() methods, restartPlaybackAfterLoss as Int
- UIDialogs: Fixed Descriptor withCondition() to accept both Boolean and () -> Boolean
- CaptchaActivity.kt: Restored from git history with WebView-based captcha solving
- JSSource.kt: Added TYPE_HLS/TYPE_DASH aliases, IV8PluginConfigStub, fromV8Video/
  fromV8Audio methods, fixed getUnderlyingObject() to return V8ValueObject
- VideoHelperExtensions.kt: getHttpDataSourceFactory() returns DefaultHttpDataSource.Factory
- StateCasting.kt: New stub file
- VideoDownloadExtensions.kt: Cleaned up duplicate imports
- strings.xml: Added missing strings (ok, copy, close, close_capitalized,
  please_enter_the_captcha_and_close_when_finished, failed_to_parse_newpipe_subscriptions)
- activity_captcha.xml: Restored from git history
- Multiple type fixes: Double->Float casts, String?->String defaults, Int->String casts

Reduction: 121 errors -> 0 errors. Build now successful.

+++ app/src/main/java/com/futo/platformplayer/states/StateDownloads.kt
@@ -510,7 +510,7 @@ class StateDownloads {
                                 val file = export.export(context, { progress ->
                                     val now = System.currentTimeMillis();
                                     if (lastNotifyTime == -1L || now - lastNotifyTime > 100) {
-                                        it.setProgress(progress);
+                                        it.setProgress(progress.toFloat());
                                         lastNotifyTime = now;
                                     }
                                 }, root);
@@ -545,7 +545,7 @@ class StateDownloads {
                         val now = System.currentTimeMillis();
                         if (lastNotifyTime == -1L || now - lastNotifyTime > 100) {
                             StateApp.instance.scopeOrNull?.launch(Dispatchers.Main) {
-                                it.setProgress(progress);
+                                it.setProgress(progress.toFloat());
                             }
                             lastNotifyTime = now;
                         }
